подключил данные с сервака, начал восстанавливать всё что отвалилось
